make codec generic & reuseable
