+++ dit-core/src/lib.rs
@@ -2,5 +2,6 @@
 #![warn(missing_debug_implementations)]
 #![deny(rustdoc::broken_intra_doc_links)]
 
+mod codec;
 pub mod daemon;
 pub mod peer;
